Top page work started, little refactoring in pages
