+++ src/pages/AboutPage/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../../components/Header';
+import Footer from '../../components/Footer';
+
+const Aboutpage = () => {
+  return (
+    <>
+      <Header />
+      <Footer />
+    </>
+  );
+};
+
+export default Aboutpage;
