--- src/pages/Aboutpage.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-
-const Aboutpage = () => {
-  return (
-    <>
-      <Header />
-      <Footer />
-    </>
-  );
-};
-
-export { Aboutpage };
