+++ src/pages/ServicePage/index.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Header from '../../components/Header';
+import Footer from '../../components/Footer';
+import Servises from '../../components/MainServises';
+
+const ServicePage = () => {
+  return (
+    <>
+      <Header />
+      <Servises />
+      <Footer />
+    </>
+  );
+};
+
+export default ServicePage;
