--- src/pages/Services.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import Servises from "../components/MainServises";
-
-const Services = () => {
-  return (
-    <>
-      <Header />
-      <Servises />
-      <Footer />
-    </>
-  );
-};
-
-export { Services };
