+++ src/pages/TopPage/styles.module.css
@@ -1,0 +1,4 @@
+.mainPageWrapper {
+    height: 100px;
+    width: 100px;
+}
