--- src/pages/Toppage.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-
-const Toppage = () => {
-  return (
-    <>
-      <Header />
-      <Footer />
-    </>
-  );
-};
-
-export { Toppage };
